requirements.txt
django
psycopg2==2.7
djangorestframework
markdown
django-filter
django-cors-headers
python_slugify==1.2.4